--- core/market/decentralized/src/matcher/mod.rs
@@ -1,5 +1,0 @@
-mod matcher;
-mod resolver;
-
-pub use matcher::Matcher;
-pub use matcher::{DemandError, MatcherError, MatcherInitError, OfferError};
