+++ core/market/decentralized/tests/test_offer_broadcast.rs
@@ -40,7 +40,7 @@ mod tests {
 
         let offer = Offer::new(json!({}), "()".to_string());
         let subscription_id = market1.subscribe_offer(&offer, &identity1).await?;
-        let offer = market1.matcher.get_offer(&subscription_id).await?;
+        let offer = market1.store.get_offer(&subscription_id).await?;
         assert!(offer.is_some());
 
         // Expect, that Offer will appear on other nodes.
@@ -48,8 +48,8 @@ mod tests {
         let market3: Arc<MarketService> = network.get_market("Node-3");
         wait_for_bcast(1000, &market2, &subscription_id, true).await?;
         // TODO: strip insertion ts from comparsion
-        assert_eq!(offer, market2.matcher.get_offer(&subscription_id).await?);
-        assert_eq!(offer, market3.matcher.get_offer(&subscription_id).await?);
+        assert_eq!(offer, market2.store.get_offer(&subscription_id).await?);
+        assert_eq!(offer, market3.store.get_offer(&subscription_id).await?);
 
         // Unsubscribe Offer. Wait some delay for propagation.
         market1
@@ -58,8 +58,8 @@ mod tests {
 
         // Expect, that Offer will disappear on other nodes.
         wait_for_bcast(1000, &market2, &subscription_id, false).await?;
-        assert!(market2.matcher.get_offer(&subscription_id).await?.is_none());
-        assert!(market3.matcher.get_offer(&subscription_id).await?.is_none());
+        assert!(market2.store.get_offer(&subscription_id).await?.is_none());
+        assert!(market3.store.get_offer(&subscription_id).await?.is_none());
 
         Ok(())
     }
@@ -88,20 +88,16 @@ mod tests {
         let identity2 = network.get_default_id("Node-2");
 
         // Prepare Offer with subscription id changed to invalid.
-        let invalid_subscription_id = SubscriptionId::from_str("c76161077d0343ab85ac986eb5f6ea38-edb0016d9f8bafb54540da34f05a8d510de8114488f23916276bdead05509a53")?;
+        let invalid_id = SubscriptionId::from_str("c76161077d0343ab85ac986eb5f6ea38-edb0016d9f8bafb54540da34f05a8d510de8114488f23916276bdead05509a53")?;
         let offer = example_offer();
         let mut offer = ModelOffer::from_new(&offer, &identity2);
-        offer.id = invalid_subscription_id.clone();
+        offer.id = invalid_id.clone();
 
         // Offer should be propagated to market1, but he should reject it.
         discovery2.broadcast_offer(offer).await?;
         tokio::time::delay_for(Duration::from_millis(50)).await;
 
-        assert!(market1
-            .matcher
-            .get_offer(&invalid_subscription_id)
-            .await?
-            .is_none());
+        assert!(market1.store.get_offer(&invalid_id).await?.is_none());
         Ok(())
     }
 
@@ -127,24 +123,24 @@ mod tests {
         let subscription_id = market1
             .subscribe_offer(&example_offer(), &identity1)
             .await?;
-        let offer = market1.matcher.get_offer(&subscription_id).await?;
+        let offer = market1.store.get_offer(&subscription_id).await?;
         assert!(offer.is_some());
 
         // Expect, that Offer will appear on other nodes.
         let market2: Arc<MarketService> = network.get_market("Node-2");
         wait_for_bcast(1000, &market2, &subscription_id, true).await?;
-        assert_eq!(offer, market2.matcher.get_offer(&subscription_id).await?);
+        assert_eq!(offer, market2.store.get_offer(&subscription_id).await?);
 
         // Unsubscribe Offer. It should be unsubscribed on all Nodes and removed from
         // database on Node-2, since it's foreign Offer.
         market1
             .unsubscribe_offer(&subscription_id, &identity1)
             .await?;
-        assert!(market1.matcher.get_offer(&subscription_id).await?.is_none());
+        assert!(market1.store.get_offer(&subscription_id).await?.is_none());
 
         // Expect, that Offer will disappear on other nodes.
         wait_for_bcast(1000, &market2, &subscription_id, false).await?;
-        assert!(market2.matcher.get_offer(&subscription_id).await?.is_none());
+        assert!(market2.store.get_offer(&subscription_id).await?.is_none());
 
         // Send the same Offer using Discovery interface directly.
         // Number of returning Offers will be counted.
